refactor: move out the duplication check in it's own class
